requirements.txt
pyecharts == 1.9.1
pyqt6-sip
requests
pandas
numpy
